+++ src/core/index.ts
@@ -12,6 +12,7 @@ export * from "./Object3D.js";
 export * from "./Scene.js";
 export * from "./Skybox.js";
 export * from "./SmallWorld.js";
+export * from "./Sprite.js";
 export * from "./cameras/strategies/IsometricStrategy.js";
 export * from "./colors/index.js";
 export * from "./events/index.js";